compressHome.sh may be fixed but still remains untested :3 Be careful

+++ bash/compressHome.sh
@@ -9,7 +9,6 @@ now=$(date +"%m_%d_%Y_%H_%M")
 
 stageName="HomeBackup_$now"
 
-
 # create the stage
 echo "creating stage $HOME/$stageName/"
 mkdir -p "$HOME/$stageName/"
@@ -20,22 +19,66 @@ if [ -d "$HOME/Archives/" ]; then
 else
     echo "No directory $HOME/Archives/ skipping..."
 fi
+
 if [ -d "$HOME/Downloads/" ]; then
     cp -R "$HOME/Downloads/" "$HOME/$stageName/"
 else
     echo "No directory $HOME/Downloads/ skipping..."
 fi
 
-cp -R "$HOME/Documents/" "$HOME/$stageName/"
-cp -R "$HOME/.font_backups/" "$HOME/$stageName/"
-cp -R "$HOME/Music/" "$HOME/$stageName/"
-cp -R "$HOME/Pictures/" "$HOME/$stageName/"
-cp -R "$HOME/Videos/" "$HOME/$stageName/"
-cp -R "$HOME/.bash_profile" "$HOME/$stageName/"
-cp -R "$HOME/.gitconfig/" "$HOME/$stageName/"
-cp -R "$HOME/.nv/" "$HOME/$stageName/"
-cp -R "$HOME/.nvidia-settings-rc" "$HOME/$stageName/"
+if [ -d "$HOME/Documents/" ]; then
+    cp -R "$HOME/Documents/" "$HOME/$stageName/"
+else
+    echo "No directory $HOME/Documents/ skipping..."
+fi
 
+if [ -d "$HOME/.font_backups/" ]; then
+    cp -R "$HOME/.font_backups/" "$HOME/$stageName/"
+else
+    echo "No directory $HOME/.font_backups/ skipping..."
+fi
+
+if [ -d "$HOME/Music/" ]; then
+    cp -R "$HOME/Music/" "$HOME/$stageName/"
+else
+    echo "No directory $HOME/Music/ skipping..."
+fi
+
+if [ -d "$HOME/Pictures/" ]; then
+    cp -R "$HOME/Pictures/" "$HOME/$stageName/"
+else
+    echo "No directory $HOME/Pictures/ skipping..."
+fi
+
+if [ -d "$HOME/Videos/" ]; then
+    cp -R "$HOME/Videos/" "$HOME/$stageName/"
+else
+    echo "No directory $HOME/Videos/ skipping..."
+fi
+
+if [ -e "$HOME/.bashrc" ]; then
+    cp -R "$HOME/.bashrc" "$HOME/$stageName/"
+else 
+    echo "No $HOME/.bashrc skipping..."
+fi
+
+if [ -e "$HOME/.gitconfig/" ]; then
+    cp -R "$HOME/.gitconfig/" "$HOME/$stageName/"
+else
+    echo "No $HOME/.gitconfig/ skipping..."
+fi
+
+if [ -e "$HOME/.nv/" ]; then
+    cp -R "$HOME/.nv/" "$HOME/$stageName/"
+else
+    echo "No $HOME/.nv/ skipping..."
+fi
+
+if [ -e "$HOME/.nv/" ]; then
+    cp -R "$HOME/.nvidia-settings-rc" "$HOME/$stageName/"
+else
+    echo "No $HOME/.nv/ skipping..."
+fi
 
 # compress stage into archive
 echo "beginning compression"
